WIP: Refactor Wrapper Generation Script
